requirements.txt
langchain
pinecone-client
openai
unstructured
streamlit